+++ circe/src/main/scala/org/http4s/circe/CirceInstances.scala
@@ -3,6 +3,7 @@ package circe
 
 import io.circe.{Encoder, Decoder, Json, Printer}
 import io.circe.jawn.CirceSupportParser.facade
+import org.http4s.batteries._
 import org.http4s.headers.`Content-Type`
 
 // Originally based on ArgonautInstances
@@ -18,14 +19,38 @@ trait CirceInstances {
       )
     }
 
-  implicit val jsonEncoder: EntityEncoder[Json] =
+  protected def defaultPrinter: Printer
+
+  implicit def jsonEncoder: EntityEncoder[Json] =
+    jsonEncoderWithPrinter(defaultPrinter)
+
+  def jsonEncoderWithPrinter(printer: Printer): EntityEncoder[Json] =
     EntityEncoder[String].contramap[Json] { json =>
       // Comment from ArgonautInstances (which this code is based on):
       // TODO naive implementation materializes to a String.
       // See https://github.com/non/jawn/issues/6#issuecomment-65018736
-      Printer.noSpaces.pretty(json)
+      printer.pretty(json)
     }.withContentType(`Content-Type`(MediaType.`application/json`))
 
   def jsonEncoderOf[A](implicit encoder: Encoder[A]): EntityEncoder[A] =
-    jsonEncoder.contramap[A](encoder.apply)
+    jsonEncoderWithPrinterOf(defaultPrinter)
+
+  def jsonEncoderWithPrinterOf[A](printer: Printer)(implicit encoder: Encoder[A]): EntityEncoder[A] =
+    jsonEncoderWithPrinter(printer).contramap[A](encoder.apply)
+
+  implicit val encodeUri: Encoder[Uri] =
+    Encoder.encodeString.contramap[Uri](_.toString)
+
+  implicit val decodeUri: Decoder[Uri] =
+    Decoder.decodeString.emap { str =>
+      Uri.fromString(str).leftMap(_ => "Uri")
+    }
+}
+
+object CirceInstances {
+  def withPrinter(p: Printer): CirceInstances = {
+    new CirceInstances {
+      def defaultPrinter: Printer = p
+    }
+  }
 }