+++ circe/src/test/scala/org/http4s/circe/CirceSpec.scala
@@ -1,12 +1,13 @@
 package org.http4s
-package circe
+package circe.test // Get out of circe package so we can import custom instances
 
 import java.nio.charset.StandardCharsets
 
 import io.circe._
+import io.circe.syntax._
+import org.http4s.circe._
 import org.http4s.headers.`Content-Type`
 import org.http4s.jawn.JawnDecodeSupportSpec
-import org.http4s.EntityEncoderSpec.writeToString
 import Status.Ok
 import org.specs2.specification.core.Fragment
 
@@ -18,8 +19,10 @@ class CirceSpec extends JawnDecodeSupportSpec[Json] {
   val foo = Foo(42)
   // Beware of possible conflicting shapeless versions if using the circe-generic module
   // to derive these.
-  implicit val FooDecoder = Decoder.instance(_.get("bar")(Decoder[Int]).map(Foo))
-  implicit val FooEncoder = Encoder.instance[Foo](foo => Json.obj("bar" -> Encoder[Int].apply(foo.bar)))
+  implicit val FooDecoder: Decoder[Foo] =
+    Decoder.forProduct1("bar")(Foo.apply)
+  implicit val FooEncoder: Encoder[Foo] =
+    Encoder.forProduct1("bar")(foo => (foo.bar))
 
   "json encoder" should {
     val json = Json.obj("test" -> Json.fromString("CirceSupport"))
@@ -31,6 +34,22 @@ class CirceSpec extends JawnDecodeSupportSpec[Json] {
     "write compact JSON" in {
       writeToString(json) must_== ("""{"test":"CirceSupport"}""")
     }
+
+    "write JSON according to custom encoders" in {
+      val custom = CirceInstances.withPrinter(Printer.spaces2)
+      import custom._
+      writeToString(json) must_== (
+        """{
+          |  "test" : "CirceSupport"
+          |}""".stripMargin)
+    }
+
+    "write JSON according to explicit printer" in {
+      writeToString(json)(jsonEncoderWithPrinter(Printer.spaces2)) must_== (
+        """{
+          |  "test" : "CirceSupport"
+          |}""".stripMargin)
+    }
   }
 
   "jsonEncoderOf" should {
@@ -40,6 +59,22 @@ class CirceSpec extends JawnDecodeSupportSpec[Json] {
 
     "write compact JSON" in {
       writeToString(foo)(jsonEncoderOf[Foo]) must_== ("""{"bar":42}""")
+    }
+
+    "write JSON according to custom encoders" in {
+      val custom = CirceInstances.withPrinter(Printer.spaces2)
+      import custom._
+      writeToString(foo)(jsonEncoderOf) must_== (
+        """{
+          |  "bar" : 42
+          |}""".stripMargin)
+    }
+
+    "write JSON according to explicit printer" in {
+      writeToString(foo)(jsonEncoderWithPrinterOf(Printer.spaces2)) must_== (
+        """{
+          |  "bar" : 42
+          |}""".stripMargin)
     }
   }
 
@@ -64,12 +99,21 @@ class CirceSpec extends JawnDecodeSupportSpec[Json] {
     // https://github.com/http4s/http4s/issues/514
     Fragment.foreach(Seq("ärgerlich", """"ärgerlich"""")) { wort =>
       sealed case class Umlaut(wort: String)
-      implicit val umlautDecoder = Decoder.instance(_.get("wort")(Decoder[String]).map(Umlaut))
+      implicit val umlautDecoder =
+        Decoder.forProduct1("wort")(Umlaut.apply)
       s"handle JSON with umlauts: $wort" >> {
         val json = Json.obj("wort" -> Json.fromString(wort))
         val result = jsonOf[Umlaut].decode(Request().withBody(json).unsafeRun, strict = true)
         result.value.unsafeRun must_== Right(Umlaut(wort))
       }
+    }
+  }
+
+  "Uri codec" should {
+    "round trip" in {
+      // TODO would benefit from Arbitrary[Uri]
+      val uri = Uri.uri("http://www.example.com/")
+      uri.asJson.as[Uri] must beRight(uri)
     }
   }
 }