+++ core/src/main/scala/org/http4s/Uri.scala
@@ -8,6 +8,7 @@ import scala.reflect.macros.whitebox.Context
 import macrocompat.bundle
 import org.http4s.Uri._
 
+import org.http4s.batteries._
 import org.http4s.parser._
 import org.http4s.util.{ Writer, Renderable, CaseInsensitiveString, UrlCodingUtils }
 import org.http4s.util.string.ToCaseInsensitiveStringSyntax
@@ -133,6 +134,15 @@ object Uri extends UriFunctions {
   /** Decodes the String to a [[Uri]] using the RFC 3986 uri decoding specification */
   def fromString(s: String): ParseResult[Uri] = new RequestUriParser(s, StandardCharsets.UTF_8).Uri
     .run()(parseResultDeliveryScheme)
+
+  /** Parses a String to a [[Uri]] according to RFC 3986.  If decoding
+   *  fails, throws a [[ParseFailure]].
+   *
+   *  For totality, call [[#fromString]].  For compile-time
+   *  verification of literals, call [[#uri]].
+   */
+  def unsafeFromString(s: String): Uri =
+    fromString(s).valueOr(throw _)
 
   /** Decodes the String to a [[Uri]] using the RFC 7230 section 5.3 uri decoding specification */
   def requestTarget(s: String): ParseResult[Uri] = new RequestUriParser(s, StandardCharsets.UTF_8).RequestUri