+++ json4s-native/src/main/scala/org/http4s/json4s/native/Json4sNativeInstances.scala
@@ -2,9 +2,9 @@ package org.http4s
 package json4s
 package native
 
-import org.json4s.native.JsonMethods
-
-import scala.text.Document
+import org.json4s.{JsonFormat, JValue, Writer}
+import org.json4s.JsonAST.JString
+import org.json4s.native.{Document, JsonMethods}
 
 trait Json4sNativeInstances extends Json4sInstances[Document] {
   override protected def jsonMethods: org.json4s.JsonMethods[Document] = JsonMethods