+++ server/src/main/scala/org/http4s/server/middleware/authentication/package.scala
@@ -6,10 +6,6 @@ import fs2._
 import org.http4s.headers._
 
 package object authentication {
-  // A function mapping (realm, username) to password, None if no password
-  // exists for that (realm, username) pair.
-  type AuthenticationStore = (String, String) => Task[Option[String]]
-
   def challenged[A](challenge: Service[Request, Either[Challenge, AuthedRequest[A]]])
                    (service: AuthedService[A]): HttpService =
     Service.lift { req =>