+++ testing/src/main/scala/org/http4s/testing/TaskMatchers.scala
@@ -1,7 +1,7 @@
 /* Derived from https://raw.githubusercontent.com/etorreborre/specs2/c0cbfc71390b644db1a5deeedc099f74a237ebde/matcher-extra/src/main/scala-scalaz-7.0.x/org/specs2/matcher/TaskMatchers.scala
  * License: https://raw.githubusercontent.com/etorreborre/specs2/master/LICENSE.txt
  */
-package org.http4s
+package org.http4s.testing
 
 import scala.concurrent.TimeoutException
 import scala.concurrent.duration.FiniteDuration