+++ tests/src/test/scala/org/http4s/multipart/MultipartSpec.scala
@@ -93,7 +93,7 @@ class MultipartSpec extends Specification with DisjunctionMatchers {
 
   def encodeAndDecodeMultipartWithBinaryFormData = {
 
-    val path       = "./core/src/test/resources/Animated_PNG_example_bouncing_beach_ball.png"
+    val path       = "./tests/src/test/resources/Animated_PNG_example_bouncing_beach_ball.png"
     val file       = new File(path)
     
     val field1     = Part.formData("field1", "Text_Field_1")